Added calendar and timer (needs to be improved)
For the calendar tab, I added the agenda component and made it so that there is an add button besides every task (need to make it so that there is only one for each day) and the tasks are autogenerated (needs to find a way to allow user to input tasks).

For the timer function, I made ti so that it will be able to fun in the back ground. But there are a few bugs like the timer will only change if you press reset timer button and the start button doesn't change when timer goes off, also the timer only works if you keep the time in the input thing, if you delete it then the time will go to zero.